refact: [NETWORK] multiple network configs
