+++ FloconDesktop/composeApp/schemas/io.github.openflocon.flocondesktop.common.db.AppDatabase/39.json
@@ -2,7 +2,7 @@
   "formatVersion": 1,
   "database": {
     "version": 39,
-    "identityHash": "121dfc4cfd03bb71ddd3d8af08dd8939",
+    "identityHash": "2f04f00102ac26b72949768d76f33930",
     "entities": [
       {
         "tableName": "FloconNetworkCallEntity",
@@ -998,8 +998,20 @@
       },
       {
         "tableName": "BadQualityConfigEntity",
-        "createSql": "CREATE TABLE IF NOT EXISTS `${TABLE_NAME}` (`deviceId` TEXT NOT NULL, `packageName` TEXT NOT NULL, `isEnabled` INTEGER NOT NULL, `errorProbability` REAL NOT NULL, `errors` TEXT NOT NULL, `triggerProbability` REAL NOT NULL, `minLatencyMs` INTEGER NOT NULL, `maxLatencyMs` INTEGER NOT NULL, PRIMARY KEY(`deviceId`, `packageName`))",
+        "createSql": "CREATE TABLE IF NOT EXISTS `${TABLE_NAME}` (`id` TEXT NOT NULL, `name` TEXT NOT NULL, `deviceId` TEXT NOT NULL, `packageName` TEXT NOT NULL, `isEnabled` INTEGER NOT NULL, `errorProbability` REAL NOT NULL, `errors` TEXT NOT NULL, `triggerProbability` REAL NOT NULL, `minLatencyMs` INTEGER NOT NULL, `maxLatencyMs` INTEGER NOT NULL, PRIMARY KEY(`id`))",
         "fields": [
+          {
+            "fieldPath": "id",
+            "columnName": "id",
+            "affinity": "TEXT",
+            "notNull": true
+          },
+          {
+            "fieldPath": "name",
+            "columnName": "name",
+            "affinity": "TEXT",
+            "notNull": true
+          },
           {
             "fieldPath": "deviceId",
             "columnName": "deviceId",
@@ -1052,15 +1064,26 @@
         "primaryKey": {
           "autoGenerate": false,
           "columnNames": [
-            "deviceId",
-            "packageName"
+            "id"
           ]
-        }
+        },
+        "indices": [
+          {
+            "name": "index_BadQualityConfigEntity_deviceId_packageName",
+            "unique": false,
+            "columnNames": [
+              "deviceId",
+              "packageName"
+            ],
+            "orders": [],
+            "createSql": "CREATE INDEX IF NOT EXISTS `index_BadQualityConfigEntity_deviceId_packageName` ON `${TABLE_NAME}` (`deviceId`, `packageName`)"
+          }
+        ]
       }
     ],
     "setupQueries": [
       "CREATE TABLE IF NOT EXISTS room_master_table (id INTEGER PRIMARY KEY,identity_hash TEXT)",
-      "INSERT OR REPLACE INTO room_master_table (id,identity_hash) VALUES(42, '121dfc4cfd03bb71ddd3d8af08dd8939')"
+      "INSERT OR REPLACE INTO room_master_table (id,identity_hash) VALUES(42, '2f04f00102ac26b72949768d76f33930')"
     ]
   }
 }
