+++ FloconDesktop/composeApp/src/commonMain/kotlin/io/github/openflocon/flocondesktop/features/network/badquality/list/mapper/BadQualityMapper.kt
@@ -1,0 +1,10 @@
+package io.github.openflocon.flocondesktop.features.network.badquality.list.mapper
+
+import io.github.openflocon.domain.network.models.BadQualityConfigDomainModel
+import io.github.openflocon.flocondesktop.features.network.badquality.list.model.NetworkBadQualityLineUiModel
+
+fun BadQualityConfigDomainModel.toLineUi() = NetworkBadQualityLineUiModel(
+    id = id,
+    name = name,
+    isEnabled = isEnabled,
+)
