+++ FloconDesktop/composeApp/src/commonMain/kotlin/io/github/openflocon/flocondesktop/features/network/mock/edition/model/MockNetworkMethodUi.kt
@@ -1,0 +1,10 @@
+package io.github.openflocon.flocondesktop.features.network.mock.edition.model
+
+enum class MockNetworkMethodUi(val text: String) {
+    GET("GET"),
+    POST("POST"),
+    PUT("PUT"),
+    DELETE("DELETE"),
+    PATCH("PATCH"),
+    ALL("*"),
+}
