--- FloconDesktop/composeApp/src/commonMain/kotlin/io/github/openflocon/flocondesktop/features/network/model/mocks/MockEditionWindowUiModel.kt
@@ -1,12 +1,0 @@
-package io.github.openflocon.flocondesktop.features.network.model.mocks
-
-import androidx.compose.runtime.Immutable
-import kotlin.uuid.ExperimentalUuidApi
-import kotlin.uuid.Uuid
-
-@OptIn(ExperimentalUuidApi::class)
-@Immutable
-data class MockEditionWindowUiModel(
-    val selectedMockUiModel: SelectedMockUiModel,
-    val windowInstanceId: String = Uuid.random().toString(),
-)
