--- FloconDesktop/composeApp/src/commonMain/kotlin/io/github/openflocon/flocondesktop/features/network/model/mocks/MockNetworkLineUiModel.kt
@@ -1,18 +1,0 @@
-package io.github.openflocon.flocondesktop.features.network.model.mocks
-
-import androidx.compose.runtime.Immutable
-
-@Immutable
-data class MockNetworkLineUiModel(
-    val id: String,
-    val isEnabled: Boolean,
-    val urlPattern: String,
-    val method: MockNetworkMethodUi,
-)
-
-fun previewMockNetworkLineUiModel(method: MockNetworkMethodUi = MockNetworkMethodUi.GET) = MockNetworkLineUiModel(
-    id = "1",
-    isEnabled = true,
-    urlPattern = "http://.*youtube.*v=.*",
-    method = method,
-)
