--- FloconDesktop/composeApp/src/commonMain/kotlin/io/github/openflocon/flocondesktop/features/network/model/mocks/MockNetworkMethodUi.kt
@@ -1,10 +1,0 @@
-package io.github.openflocon.flocondesktop.features.network.model.mocks
-
-enum class MockNetworkMethodUi(val text: String) {
-    GET("GET"),
-    POST("POST"),
-    PUT("PUT"),
-    DELETE("DELETE"),
-    PATCH("PATCH"),
-    ALL("*"),
-}
