--- FloconDesktop/composeApp/src/commonMain/kotlin/io/github/openflocon/flocondesktop/features/network/model/mocks/SelectedMockUiModel.kt
@@ -1,12 +1,0 @@
-package io.github.openflocon.flocondesktop.features.network.model.mocks
-
-import androidx.compose.runtime.Immutable
-
-@Immutable
-sealed interface SelectedMockUiModel {
-    @Immutable
-    data object Creation : SelectedMockUiModel
-
-    @Immutable
-    data class Edition(val existing: MockNetworkUiModel) : SelectedMockUiModel
-}
