+++ FloconDesktop/composeApp/src/commonMain/kotlin/io/github/openflocon/flocondesktop/features/network/view/NetworkScreen.kt
@@ -33,10 +33,10 @@ import io.github.openflocon.flocondesktop.features.network.model.NetworkItemView
 import io.github.openflocon.flocondesktop.features.network.model.previewGraphQlItemViewState
 import io.github.openflocon.flocondesktop.features.network.model.previewNetworkItemViewState
 import io.github.openflocon.flocondesktop.features.network.previewNetworkUiState
-import io.github.openflocon.flocondesktop.features.network.view.badquality.BadNetworkQualityWindow
+import io.github.openflocon.flocondesktop.features.network.badquality.list.BadNetworkQualityWindow
 import io.github.openflocon.flocondesktop.features.network.view.header.NetworkFilter
 import io.github.openflocon.flocondesktop.features.network.view.header.NetworkItemHeaderView
-import io.github.openflocon.flocondesktop.features.network.view.mocks.NetworkMocksWindow
+import io.github.openflocon.flocondesktop.features.network.mock.list.view.NetworkMocksWindow
 import io.github.openflocon.library.designsystem.FloconTheme
 import io.github.openflocon.library.designsystem.components.FloconPanel
 import io.github.openflocon.library.designsystem.components.FloconSurface
