--- FloconDesktop/composeApp/src/commonMain/kotlin/io/github/openflocon/flocondesktop/features/network/view/mocks/MockNetworkLabelView.kt
@@ -1,21 +1,0 @@
-package io.github.openflocon.flocondesktop.features.network.view.mocks
-
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.dp
-import io.github.openflocon.library.designsystem.FloconTheme
-
-@Composable
-internal fun MockNetworkLabelView(label: String) {
-    Text(
-        label,
-        modifier = Modifier.padding(start = 4.dp),
-        color = FloconTheme.colorPalette.onSurface,
-        style = FloconTheme.typography.bodyMedium.copy(
-            fontWeight = FontWeight.Thin,
-        ),
-    )
-}
