+++ FloconDesktop/data/core/src/commonMain/kotlin/io/github/openflocon/data/core/network/datasource/NetworkQualityLocalDataSource.kt
@@ -2,6 +2,7 @@ package io.github.openflocon.data.core.network.datasource
 
 import io.github.openflocon.domain.device.models.DeviceIdAndPackageNameDomainModel
 import io.github.openflocon.domain.network.models.BadQualityConfigDomainModel
+import io.github.openflocon.domain.network.models.BadQualityConfigId
 import kotlinx.coroutines.flow.Flow
 
 interface NetworkQualityLocalDataSource {
@@ -10,15 +11,32 @@ interface NetworkQualityLocalDataSource {
         config: BadQualityConfigDomainModel
     )
 
-    suspend fun getNetworkQuality(deviceIdAndPackageName: DeviceIdAndPackageNameDomainModel) : BadQualityConfigDomainModel?
+    suspend fun getNetworkQuality(
+        deviceIdAndPackageName: DeviceIdAndPackageNameDomainModel,
+        configId: BadQualityConfigId,
+    ) : BadQualityConfigDomainModel?
 
     fun observe(
         deviceIdAndPackageName: DeviceIdAndPackageNameDomainModel,
+        configId: BadQualityConfigId,
     ): Flow<BadQualityConfigDomainModel?>
 
-    suspend fun updateIsEnabled(
+    suspend fun delete(
         deviceIdAndPackageName: DeviceIdAndPackageNameDomainModel,
-        isEnabled: Boolean,
+        configId: BadQualityConfigId,
     )
+
+    fun observeAll(
+        deviceIdAndPackageName: DeviceIdAndPackageNameDomainModel,
+    ): Flow<List<BadQualityConfigDomainModel>>
+
+    suspend fun setEnabledConfig(
+        deviceIdAndPackageName: DeviceIdAndPackageNameDomainModel,
+        configId: BadQualityConfigId?,
+    )
+
+    suspend fun getTheOnlyEnabledNetworkQuality(
+        deviceIdAndPackageName: DeviceIdAndPackageNameDomainModel
+    ): BadQualityConfigDomainModel?
 
 }
