+++ FloconDesktop/data/local/src/commonMain/kotlin/io/github/openflocon/data/local/network/models/badquality/BadQualityConfigEntity.kt
@@ -2,10 +2,19 @@ package io.github.openflocon.data.local.network.models.badquality
 
 import androidx.room.Embedded
 import androidx.room.Entity
+import androidx.room.Index
+import androidx.room.PrimaryKey
 
-@Entity(primaryKeys = ["deviceId", "packageName"])
+@Entity(
+    indices = [
+        Index(value = ["deviceId", "packageName"]),
+    ],
+)
 data class BadQualityConfigEntity(
+    @PrimaryKey val id: String,
+    val name: String,
     val deviceId: String,
+    val createdAt: Long,
     val packageName: String,
     val isEnabled: Boolean,
     @Embedded val latency: LatencyConfigEmbedded,
