+++ FloconDesktop/domain/src/commonMain/kotlin/io/github/openflocon/domain/network/DI.kt
@@ -9,10 +9,12 @@ import io.github.openflocon.domain.network.usecase.RemoveHttpRequestUseCase
 import io.github.openflocon.domain.network.usecase.RemoveHttpRequestsBeforeUseCase
 import io.github.openflocon.domain.network.usecase.ResetCurrentDeviceHttpRequestsUseCase
 import io.github.openflocon.domain.network.usecase.UpdateNetworkFilterUseCase
+import io.github.openflocon.domain.network.usecase.badquality.DeleteBadQualityUseCase
+import io.github.openflocon.domain.network.usecase.badquality.ObserveAllNetworkBadQualitiesUseCase
 import io.github.openflocon.domain.network.usecase.badquality.ObserveNetworkBadQualityUseCase
 import io.github.openflocon.domain.network.usecase.badquality.SaveNetworkBadQualityUseCase
 import io.github.openflocon.domain.network.usecase.badquality.SetupNetworkBadQualityUseCase
-import io.github.openflocon.domain.network.usecase.badquality.UpdateNetworkBadQualityIsEnabledUseCase
+import io.github.openflocon.domain.network.usecase.badquality.SetNetworkBadQualityEnabledConfigUseCase
 import io.github.openflocon.domain.network.usecase.mocks.AddNetworkMocksUseCase
 import io.github.openflocon.domain.network.usecase.mocks.DeleteNetworkMocksUseCase
 import io.github.openflocon.domain.network.usecase.mocks.GenerateNetworkMockFromNetworkCallUseCase
@@ -45,6 +47,8 @@ internal val networkModule = module {
     // bad quality
     factoryOf(::ObserveNetworkBadQualityUseCase)
     factoryOf(::SaveNetworkBadQualityUseCase)
+    factoryOf(::DeleteBadQualityUseCase)
     factoryOf(::SetupNetworkBadQualityUseCase)
-    factoryOf(::UpdateNetworkBadQualityIsEnabledUseCase)
+    factoryOf(::SetNetworkBadQualityEnabledConfigUseCase)
+    factoryOf(::ObserveAllNetworkBadQualitiesUseCase)
 }
