+++ FloconDesktop/domain/src/commonMain/kotlin/io/github/openflocon/domain/network/repository/NetworkBadQualityRepository.kt
@@ -2,6 +2,7 @@ package io.github.openflocon.domain.network.repository
 
 import io.github.openflocon.domain.device.models.DeviceIdAndPackageNameDomainModel
 import io.github.openflocon.domain.network.models.BadQualityConfigDomainModel
+import io.github.openflocon.domain.network.models.BadQualityConfigId
 import kotlinx.coroutines.flow.Flow
 
 interface NetworkBadQualityRepository {
@@ -19,14 +20,29 @@ interface NetworkBadQualityRepository {
 
     suspend fun getNetworkQuality(
         deviceIdAndPackageName: DeviceIdAndPackageNameDomainModel,
+        configId: BadQualityConfigId,
     ) : BadQualityConfigDomainModel?
+
+    suspend fun getTheOnlyEnabledNetworkQuality(
+        deviceIdAndPackageName: DeviceIdAndPackageNameDomainModel,
+    ) : BadQualityConfigDomainModel?
+
+    suspend fun deleteNetworkQuality(
+        deviceIdAndPackageName: DeviceIdAndPackageNameDomainModel,
+        configId: BadQualityConfigId,
+    )
 
     fun observeNetworkQuality(
         deviceIdAndPackageName: DeviceIdAndPackageNameDomainModel,
+        configId: BadQualityConfigId,
     ) : Flow<BadQualityConfigDomainModel?>
 
-    suspend fun setNetworkQualityIsEnabled(
+    fun observeAllNetworkQualities(
         deviceIdAndPackageName: DeviceIdAndPackageNameDomainModel,
-        isEnabled: Boolean,
+    ) : Flow<List<BadQualityConfigDomainModel>>
+
+    suspend fun setEnabledConfig(
+        deviceIdAndPackageName: DeviceIdAndPackageNameDomainModel,
+        configId: BadQualityConfigId?,
     )
 }
