+++ FloconDesktop/domain/src/commonMain/kotlin/io/github/openflocon/domain/network/usecase/badquality/ObserveAllNetworkBadQualitiesUseCase.kt
@@ -2,30 +2,24 @@ package io.github.openflocon.domain.network.usecase.badquality
 
 import io.github.openflocon.domain.device.usecase.ObserveCurrentDeviceIdAndPackageNameUseCase
 import io.github.openflocon.domain.network.models.BadQualityConfigDomainModel
-import io.github.openflocon.domain.network.models.BadQualityConfigId
-import io.github.openflocon.domain.network.models.MockNetworkDomainModel
 import io.github.openflocon.domain.network.repository.NetworkBadQualityRepository
-import io.github.openflocon.domain.network.repository.NetworkMocksRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.flowOf
 
-class ObserveNetworkBadQualityUseCase(
+class ObserveAllNetworkBadQualitiesUseCase(
     private val networkBadQualityRepository: NetworkBadQualityRepository,
     private val observeCurrentDeviceIdAndPackageNameUseCase: ObserveCurrentDeviceIdAndPackageNameUseCase,
 ) {
-    operator fun invoke(
-        configId: BadQualityConfigId,
-    ): Flow<BadQualityConfigDomainModel?> =
+    operator fun invoke(): Flow<List<BadQualityConfigDomainModel>> =
         observeCurrentDeviceIdAndPackageNameUseCase()
             .flatMapLatest { current ->
                 if (current == null) {
-                    flowOf(null)
+                    flowOf(emptyList())
                 } else {
-                    networkBadQualityRepository.observeNetworkQuality(
+                    networkBadQualityRepository.observeAllNetworkQualities(
                         deviceIdAndPackageName = current,
-                        configId = configId,
                     )
                 }
             }
