+++ FloconDesktop/domain/src/commonMain/kotlin/io/github/openflocon/domain/network/usecase/badquality/SetNetworkBadQualityEnabledConfigUseCase.kt
@@ -1,21 +1,24 @@
 package io.github.openflocon.domain.network.usecase.badquality
 
 import io.github.openflocon.domain.device.usecase.GetCurrentDeviceIdAndPackageNameUseCase
+import io.github.openflocon.domain.network.models.BadQualityConfigId
 import io.github.openflocon.domain.network.repository.NetworkBadQualityRepository
-import io.github.openflocon.domain.network.repository.NetworkMocksRepository
 
-class SetupNetworkBadQualityUseCase(
+class SetNetworkBadQualityEnabledConfigUseCase(
     private val getCurrentDeviceIdAndPackageNameUseCase: GetCurrentDeviceIdAndPackageNameUseCase,
     private val networkBadQualityRepository: NetworkBadQualityRepository,
+    private val setupNetworkBadQualityUseCase: SetupNetworkBadQualityUseCase,
 ) {
-    suspend operator fun invoke() {
+    suspend operator fun invoke(
+        configId: BadQualityConfigId?, // null to enable none
+    ) {
         getCurrentDeviceIdAndPackageNameUseCase()?.let { deviceIdAndPackageName ->
-            networkBadQualityRepository.setupBadNetworkQuality(
-                config = networkBadQualityRepository.getTheOnlyEnabledNetworkQuality(
-                    deviceIdAndPackageName = deviceIdAndPackageName,
-                ),
+            networkBadQualityRepository.setEnabledConfig(
+                configId = configId,
                 deviceIdAndPackageName = deviceIdAndPackageName,
             )
+            // then send to device
+            setupNetworkBadQualityUseCase()
         }
     }
 }
